+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     + \#54169: Require jQuery in the engine
     + \#54608: Highlight next section on detail page's sub nav click
     + \#54608: Include `on_production` in live toggle attribute check
+    + \#53749: Hide filters when no records are present
 
 ## 1.2.3